refactor(accelerator): extract server::auth (authorize_origin) from server.rs (Q2 4/n)
Move the /prove origin-authorization handler into server/auth.rs (child
module reaches parent's private json_error/ProveError/AUTH_DECISION_TIMEOUT
via super::). Behaviour-preserving: the prove_* auth tests exercise it via
/prove and stay green. Drops the now-unused AuthDecision import from server.rs.
